Implemented MenuScene class into the main menu of the game
Now, theres' a good-looking view in the main menu of the game.

This basically is the final version of the game (at least at the moment). So, tomorrow (just in a few hours) I gonna make a release version into this project and design all the "release stuff" thing.

GG